populate with a new user correction of the test margin modification in login.scss

+++ src/test/java/fr/eventmanager/dao/EventDAOTest.java
@@ -45,20 +45,20 @@ public class EventDAOTest {
         List<Event> eventsFourthPage = eventDAO.getPageEvents(4);
 
         assertEquals(eventsFirstPage.size(), 4);
-        assertEquals(eventsFirstPage.get(0).getId(), new Integer(3));
-        assertEquals(eventsFirstPage.get(3).getId(), new Integer(6));
+        assertEquals(eventsFirstPage.get(0).getId(), new Integer(6));
+        assertEquals(eventsFirstPage.get(3).getId(), new Integer(9));
 
         assertEquals(eventsSecondPage.size(), 4);
-        assertEquals(eventsSecondPage.get(0).getId(), new Integer(7));
-        assertEquals(eventsSecondPage.get(3).getId(), new Integer(10));
+        assertEquals(eventsSecondPage.get(0).getId(), new Integer(10));
+        assertEquals(eventsSecondPage.get(3).getId(), new Integer(13));
 
         assertEquals(eventsThirdPage.size(), 4);
-        assertEquals(eventsThirdPage.get(0).getId(), new Integer(11));
-        assertEquals(eventsThirdPage.get(3).getId(), new Integer(14));
+        assertEquals(eventsThirdPage.get(0).getId(), new Integer(14));
+        assertEquals(eventsThirdPage.get(3).getId(), new Integer(17));
 
         assertEquals(eventsFourthPage.size(), 3);
-        assertEquals(eventsFourthPage.get(0).getId(), new Integer(15));
-        assertEquals(eventsFourthPage.get(2).getId(), new Integer(17));
+        assertEquals(eventsFourthPage.get(0).getId(), new Integer(18));
+        assertEquals(eventsFourthPage.get(2).getId(), new Integer(20));
 
 
     }
